revert utils removal; make use of orch.utils in slurm launcher
